move remainingTime to FindingController
