+++ conf/routes
@@ -38,14 +38,8 @@ GET     /Finding/:findingIdentifier/getBrainstormingFinding         controllers.
 PUT     /Finding/:findingIdentifier/putBrainsheet                   controllers.FindingController.putBrainsheet(findingIdentifier)
 + noJwtFilter
 GET     /Finding/:findingIdentifier/startBrainstorming              controllers.FindingController.startBrainstorming(findingIdentifier)
-
-
 + noJwtFilter
-GET     /Finding/:teamIdentifier/:findingIdentifier/remainingTime   controllers.TimerController.calculateRemainingTimeOfFindingFromTeam(teamIdentifier, findingIdentifier)
-
-
-
-
+GET     /Finding/:findingIdentifier/getRemainingTime                controllers.FindingController.calculateRemainingTimeOfFinding(findingIdentifier)
